Completed MessageHandler
MessageHandler is completed, will now output logs to log file. Most of
the functionality for logging to Windows Event Viewer is there, however
this is not working at the moment due to the user requiring admin access
to create an Event Source in Windows. Need to implement this as part of
an installer.